move is_external_address to global_helpers
