+++ global_helpers/panther_msft_helpers.py
@@ -57,3 +57,39 @@ def get_target_name(event, target_type="User"):
 def azure_success(event):
     result = event.deep_get("properties", "result", default="")
     return result == "success"
+
+
+def is_external_address(address, primary_domain, onmicrosoft_domain):
+    """Check if an email address is external to the organization.
+
+    Args:
+        address (str): The email address or SMTP address to check
+        primary_domain (str): The organization's primary domain (e.g. contoso.com)
+        onmicrosoft_domain (str): The tenant domain (e.g. contoso.onmicrosoft.com)
+
+    Returns:
+        bool: True if the address is external, False if internal
+    """
+    if not address or (not primary_domain and not onmicrosoft_domain):
+        return True
+
+    # Clean up and normalize the address
+    address = address.lower()
+    if address.startswith("smtp:"):
+        address = address[5:]
+
+    # Check each address (might be multiple addresses separated by semicolon)
+    for addr in address.split(";"):
+        try:
+            domain = addr.strip().split("@")[1].lower()
+            # Skip if internal (matches onmicrosoft domain or primary domain/subdomain)
+            if (onmicrosoft_domain and domain == onmicrosoft_domain.lower()) or (
+                primary_domain
+                and (domain == primary_domain or domain.endswith("." + primary_domain))
+            ):
+                continue
+            return True
+        except (IndexError, AttributeError):
+            return True
+
+    return False
